source/data_scraping/ms_graph/requirements.txt
python-msgraph
azure-identity
msgraph-sdk